Creacion de la clase Persona de la que extiende Médico y Paciente (#7)
* Iniciacion clases

* Completado las clases Hospita,Medico,Paciente

* Completado las clases Hospital, Medico, Paciente

* Modificacion plantilla MVC

* Creacion de persona de la que extienden Medico y Paciente

---------
