+++ src/Clases/Medico.java
@@ -1,7 +1,7 @@
 package Clases;
 
-public class Medico {
-    String codMed, nomMed, codH1;
+public class Medico extends Persona{
+    String codH1;
 
     /**
      * Constructor del médico
@@ -10,8 +10,7 @@ public class Medico {
      * @param codH1 -> código del hospital donde trabaja
      */
     public Medico(String codMed, String nomMed, String codH1) {
-        this.codMed = codMed;
-        this.nomMed = nomMed;
+        super(codMed,nomMed);
         this.codH1 = codH1;
     }
 
@@ -20,28 +19,11 @@ public class Medico {
      */
     @Override
     public String toString() {
-        return  "Codigo del médico: " + codMed +
-                ", nombre: " + nomMed +
-                ", codigo del hospital: " + codH1;
+        return  "codMedico: " + codP + ", nomP: " + nomP + "codH1: " + codH1;
     }
 
     //setters y getters
-    public String getCodMed() {
-        return codMed;
-    }
-
-    public void setCodMed(String codMed) {
-        this.codMed = codMed;
-    }
-
-    public String getNomMed() {
-        return nomMed;
-    }
-
-    public void setNomMed(String nomMed) {
-        this.nomMed = nomMed;
-    }
-
+    
     public String getCodH1() {
         return codH1;
     }
